Create Graph.py
Its not complete but radar and tape measure (and multi-view theoretically works), overhead is close to being done just need to do a little more maths. Need to do a little on threading and actually decide what to do with the data from the serial function. (it did occasionally crash but haven't seen that issue today). Also mouse position for buttons need a bit of tweaking but an easy fix. Need to test with input have tested most using pythons in built function. Need a way for one thread to interact with another thread so one can pipe data to another but seems easier enough.