MODES CHANGES
Broke modes.c into a bunch of individual files that contain the non-mode
functions that were in modes. They are lsted with the same file names as
they had headers in modes. I have also given the global variables their
own file.

This shouldn't affect anything and it compiles successfully, however it
needs to be tested.